cyberhead/requirements.txt
PyYAML==5.3.1
Backtesting==0.1.7
celery==4.4.5
gitpython
